+++ src/App.tsx
@@ -1,13 +1,10 @@
 import { Box } from '@chakra-ui/react'
-import { useState } from 'react'
 import { useLocation } from 'react-router-dom'
 import ResetPassword from './components/Auth/ResetPassword'
-import Header from './components/Header'
 import PlaygroundMap from './components/PlaygroundMap'
 import useDocumentTitle from './hooks/useDocumentTitle'
 
 function App() {
-  const [showSignIn, setShowSignIn] = useState(false)
   const location = useLocation()
   const searchParams = new URLSearchParams(location.search)
   const showResetPassword = searchParams.has('reset_password')
@@ -16,18 +13,6 @@ function App() {
   return (
     <Box position="relative" h="100vh" w="100vw">
       <PlaygroundMap />
-      <Box
-        position="absolute"
-        top={4}
-        left={4}
-        right={4}
-        zIndex={1000}
-      >
-        <Header
-          showSignIn={showSignIn}
-          setShowSignIn={setShowSignIn}
-        />
-      </Box>
       {showResetPassword && (
         <Box
           position="fixed"